Changed for raw feed
Since our lord an saviour, mike, removed a lot of spam from the homepage, the original script didn't make sense anymore. To welcome new players, I started paying attention to the raw feed on LDJAM.com and found out that this script does not work there.

I changed it, so that the text "<link />" triggers the link code.